S001 : Création d’une actualité (S001-TX001-RG005). Gestion des images

+++ src/main/java/com/jeff/actualite/domain/dto/ImageDto.java
@@ -25,7 +25,7 @@ public class ImageDto {
     private String contentType;
 
     @NotNull
-    private Long taille;
+    private Integer taille;
 
     @NotNull
     private Byte[] contenu;
